# Conflicts:
#	WalletKitCore/Package.swift
#	WalletKitCore/WalletKitCoreTestsInstallConfig.sh
#	WalletKitJava/CoreNative/CMakeLists.txt
#	WalletKitJava/CoreNative/build.gradle
#	WalletKitJava/CoreNative/src/main/cpp/core
#	WalletKitJava/crypto/src/main/java/com/breadwallet/crypto/blockchaindb/apis/PagedCompletionHandler.java
#	WalletKitSwift/.gitignore
#	WalletKitSwift/InstallCoreTestsConfig.sh

+++ WalletKitCore/src/ethereum/les/msg/BREthereumMessageLES.c
@@ -943,7 +943,7 @@ messageLESGetRequestIdInternal (const BREthereumLESMessage *message) {
 extern size_t
 messageLESGetRequestId (const BREthereumLESMessage *message) {
     uint64_t reqId = messageLESGetRequestIdInternal (message);
-    if (LES_MESSAGE_NO_REQUEST_ID == reqId) return LES_MESSAGE_NO_REQUEST_ID;
+    if (LES_MESSAGE_NO_REQUEST_ID == reqId) return (size_t) LES_MESSAGE_NO_REQUEST_ID;
     assert (reqId <= (uint64_t) SIZE_MAX);
     return (size_t) reqId;
 }
